rename typing -> typings to avoid confliction with the standard library
