+++ haiku/_src/batch_norm_test.py
@@ -15,6 +15,8 @@
 # ==============================================================================
 """Tests for haiku._src.batch_norm."""
 
+from __future__ import absolute_import
+
 from absl.testing import absltest
 from haiku._src import batch_norm
 from haiku._src import test_utils
