+++ haiku/typings.py
@@ -13,23 +13,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Haiku types."""
+"""Type hints."""
 
-from typing import Any, Callable, Mapping, Text, Sequence, Union
-import jax.numpy as jnp
+from haiku._src.typings import Params
+from haiku._src.typings import State
 
-Shape = Sequence[int]
-DType = Any
-ParamName = Text
-Initializer = Callable[[Shape, DType], jnp.ndarray]
-Params = Mapping[Text, Mapping[ParamName, jnp.ndarray]]
-State = Mapping[Text, Mapping[Text, jnp.ndarray]]
-NextCreator = Callable[[ParamName, Shape, DType, Initializer], jnp.ndarray]
-ParamCreator = Callable[[NextCreator, ParamName, Shape, DType, Initializer],
-                        jnp.ndarray]
-Padding = Callable[[int], Sequence[int]]
-Paddings = Union[Padding, Sequence[Padding]]
-
-# Missing JAX types.
-PRNGKey = jnp.ndarray  # pylint: disable=invalid-name
-PRNGSeed = int
+__all__ = (
+    "Params",
+    "State",
+)
